Moved main state to App and pushed user Batman choice to team array in state.
