+++ src/components/CharacterBlock.js
@@ -16,9 +16,9 @@ class CharacterBlock extends Component {
                     <div className="selectCharacterBlock">
                         {/* ============================================ */}
                         {/* New Team: Character Select Block */}
-                        <Link to="/Search" component={Search}>
-                            <button className="selectCharacterButton button">Select Character</button>
-                        </Link>
+
+                        <button className="selectCharacterButton button">Select Character</button>
+
                     </div>
 
                     :
@@ -40,9 +40,9 @@ class CharacterBlock extends Component {
                             <li className="statsList__item">Sp: {this.props.chara.stats.spd}</li>
                             <li className="statsList__item">St: {this.props.chara.stats.str}</li>
                         </ul>
-                        <Link to="/Search">
-                            <button className="changeCharacterButton button">Change Character</button>
-                        </Link>
+
+                        <button className="changeCharacterButton button">Change Character</button>
+
                     </div>
                 }
             </div>
