+++ src/components/Search.js
@@ -63,7 +63,7 @@ class Search extends Component {
             }
         };
         console.log('Testy McTesterface');
-        // this.props.addToRoster(charObj);
+        this.props.addToTeamArray(charObj);
     }
 
     render() {
@@ -91,9 +91,9 @@ class Search extends Component {
                                         <li className="charaStat">{chara.powerstats.combat}</li>
                                     </ul>
                                     {/* can you trigger a click event and rout at the same time? */}
-                                    <Link to="/TeamSelect">
-                                        <button onClick={() => this.addToRoster(i)}>Add to Roster</button>
-                                    </Link>
+
+                                    <button onClick={() => this.addToRoster(i)}>Add to Roster</button>
+
                                 </div>
                                 {/* if an existing character is being replaced  display the old character */}
                                 {/* <img src={oldChara.image.url} alt={`${oldChara.name} Snapshot`} />
